Clase 4/Ejemplos/paf_api/requirements.txt
Django==3.1
django-extensions
django-filter
djangorestframework
gunicorn
sqlparse
whitenoise
